File modified

+++ 0x00-hello_world/6-size.c
@@ -6,10 +6,10 @@
 #include <stdio.h>
 int main(void)
 {
-	printf("Size of a char: %d bytes(s)\n", sizeof(char));
-	printf("Size of an int: %d bytes(s)\n", sizeof(int));
-	printf("Size of a long int: %d bytes(s)", sizeof(long int));
-	print("Size of a long long int: %d bytes(s)\n", sizeof(long long int));
-	printf("Size of a float: %d bytes(s)\n", sizeof(float));
+	printf("Size of a char: %ld bytes(s)\n", (long unsigned int)sizeof(char));
+	printf("Size of an int: %ld bytes(s)\n", (long unsigned int)sizeof(int));
+	printf("Size of a long int: %ld bytes(s)", (long unsigned int)sizeof(long int));
+	print("Size of a long long int: %ld bytes(s)\n", (long unsigned int)sizeof(long long int));
+	printf("Size of a float: %ld bytes(s)\n", (long unsigned int)sizeof(float));
 	return (0);
 }
